zmq-broker: cmbutil: fix segfault when listing kvs entries
If cmbutil --list-kvs was used with a non-directory entry in the kvs,
the utility would segfault due to json_object_foreachC() being used
on a non-object json object.

Instead we first test for existing a json object of type "object"
before proceeding with the foreach. If we are not dealing with
an object, then the value is simply printed.