pepe: Do not close stdio file descriptors in pepe.run
pepe.run was setting the close-on-exec flag for all open
file descriptors, and FDs 0,1,2 were accidentally included
in this list. This resulted in havoc in child processes when
these descriptors are ropened for non-stdio related code,
and then a stdio function such as printf is used.

Instead we should only set close-on-exec for FDs 3 and above.