jsspp-paper: abstract rework [Becky]
Abstract to have another editing pass after paper structure
finalized.

Becky's note pointing to 'key components':
Mention Run-Time system???  This final sentence about scheduling did
not seem to fit here as an ending sentence.  Adding in the user
productivity here allows us to drop that as a separate section.
We can focus on it more fully, along with scheduling and security,
in later write-ups.  Again, cover this in later efforts.