librdl: Fix lua_rdl_accumulator_method_push()
lua_rdl_accumulator_method_push() was using lua_rawget() which was
avoiding invocation of the Accumulator __index metamethod. This meant
that for methods "forwarded" to the underlying "destination" RDL
object, this C call was failing.

The function also failed to return an error if a function type
was not returned as the method.