modctl: ensure module loads first on master
To handle modules that may have a "master" on rank 0, load a module
on rank 0 first, then the rest in random order.

Specifically: add flux_modctl_ins() and flux_modctl_rm() functions
that call to the modctl master.  The master loads/unloads the requested
module, then updates conf.modctl.seq in the KVS.  The slaves then follow
the KVS.