Incorporate RDL's accumulator functionality
The flux_lwj_t structure now has a member that contains the resources
the job has been allocated.  When schedule_job() finds the available
resources, it effectively copies the resources from the main RDL
containing just the resources (nodes and cores right now) and assigns
them to the job.

It also tags the allocated resources in the RDL with the job ID.
Later, when the job completes, the allocated resources will remain a
part of the job record for historical purposes, but the lwj tags will
be removed from the now available resources.