rdl: ListOf should return a list not a table
The ListOf function was returning a table of Resources instead of a
list (expanded table), and this made it awkward to chain lists of
Resources together, i.e.

  children = Listof{ ... }, Resource{ ... }

was impossible.

Instead, this patch has ListOf{} return an expanded list of resources.
This means that ListOf{} must be used inside brackets in a children
list defintion, e.g.

  children = { ListOf{ .. } }

the benefit being that arbitrary Resources can *also* be added to the
list, including multiple Lists, etc.