live: add flux-mod update
Since 'flux-mod ls' just reads data out of the KVS, the idle
times are always stale.  Add an update request that increments
the kvs sequence number to trigger an update.

Note: the design should probably be changed so that a "modctl.ls"
request, handled by the master node, triggers an lsmod reduction
and waits for the result.  Caching this data in the KVS was
weak to begin with.  Come back and fix this.