live: generate topology from hello messages
Instead of only collecting a list of children who have said
hello and using that to initialize conf.live.status,
send a bit more information in the hello message:  parent rank and
list of children.  Reduce this and pick up both topology and
status on the master.  Topology then follows additional hellos,
adding parent-child relationships (but not deleting old ones).