reactor: handler must destroy request or ENOSYS
Disable code in the reactor to allow multiple request handlers
to match a request.  I'm not sure exactly how it happens, but
if a request handler failed to destroy a handled request, it
might see it one more time.

Instead, break out of the handler-matching loop and return to
the handle implementation. plugin.c will then send an ENOSYS
response for unconsumed request.