pmi-boot: add -P,--pmi-boot and -k,--k-ary options
With these options, the cmbd can be started directly by srun.

The rank and size are read from PMI.

The child URI is dynamically generated and written to the PMI KVS
under the rank.  In order to accomodate multiple cmbd's per node,
the port number is derived from the rank.

Parent rank is computed for k-ary tree, and parent URI is read
from the PMI KVS after a commit/barrier.

A minimal set of comms modules are loaded if none is specified
on the command line.