pmi-boot: rework event plumbing
Simplify event plumbing.

Rank 0 binds a ZMQ_PUB socket to the EPGM address, which may be
specified on the command line, or for PMI, can be based on the PMI
appnum (jobid).  Ranks > 0 connect a ZMQ_SUB socket to the EPGM address.

Events are published by sending a cmb.pub request.  This is forwarded
to rank 0.  On rank 0, the event is unwrapped and sent out the EPGM
socket, and is looped back to modules and cmbd internally.
Events are received via EPGM on ranks > 0, and solely by loopback
on rank = 0.